[TCM] Fix representation of hwloc mask in the test system
The use of std::string results in the additional new line symbol when printing the mask on the
console, which makes its representation poor. Additionally, the size of the mask was computed
incorrectly, since 'hwloc_bitmap_weight' function returns the number of turned ON bits, instead of
the number of characters needed to represent the mask in the text form.

Signed-off-by: Fedotov, Aleksei <aleksei.fedotov@intel.com>